# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "eCommerce",
    "description": "Documentation for eCommerce Exam Project"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Register a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Signup"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login with username or email and password",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/auth/": {
      "put": {
        "tags": [
          "Auth"
        ],
        "description": "Changing role value on a user. Only Admin can change this",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUserRole"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Get all brands in database",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Create brand. Brand can be anything execpt for a number",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateBrand"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Updart brand name using brand name as the key",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateBrand"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "Delete a brand, if its not active on any products",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DeleteBrand"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/admin/brands": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get a list of all brands with crud functionality",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
    "Login": {
      "type": "object",
      "properties": {
        "emailOrUsername": {
          "type": "string",
          "example": "Admin"
        },
        "password": {
          "type": "string",
          "example": "P@ssword2023"
        }
      },
      "required": [
        "emailOrUsername",
        "password"
      ]
    },
    "Signup": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Admin"
        },
        "lastName": {
          "type": "string",
          "example": "Support"
        },
        "userName": {
          "type": "string",
          "example": "Admin"
        },
        "email": {
          "type": "string",
          "example": "Admin@noroff.no"
        },
        "password": {
          "type": "string",
          "example": "P@ssword2023"
        },
        "address": {
          "type": "string",
          "example": "Online"
        },
        "telephoneNumber": {
          "type": "string",
          "example": "911"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "userName",
        "email",
        "password",
        "address",
        "telephoneNumber"
      ]
    },
    "UpdateUserRole": {
      "type": "object",
      "properties": {
        "newRole": {
          "type": "string",
          "example": "1"
        },
        "id": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "newRole",
        "id"
      ]
    },
    "CreateBrand": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string",
          "example": "SwaggerBrandTest"
        }
      },
      "required": [
        "brand"
      ]
    },
    "UpdateBrand": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        },
        "updatedBrand": {
          "type": "string",
          "example": "NewBrandName"
        }
      },
      "required": [
        "id",
        "updatedBrand"
      ]
    },
    "DeleteBrand": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string",
          "example": "SwaggerBrandTest"
        }
      },
      "required": [
        "brand"
      ]
    },
...
  }
}